Cleanups
Better comments and refactorings that move some things around
so that less modules depend on Inliner.
